% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cod_gr}
\alias{cod_gr}
\title{cod_gr dataset}
\format{A dataframe with 3 data columns and 1 metadata column.
\describe{
\item{seqnames}{chromosome names (chr21)}
\item{ranges}{ranges}{chromosome ranges on the genome (10906201-11029719)}
\item{strand}{specific strand of the genomic location (+,-,*)}
}}
\usage{
cod_gr
}
\description{
A subset of gencode_gr extracted as:
cod_gr <- subset(gencode_gr, biotype == 'protein_coding')
}
\keyword{datasets}
