% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReadthrough.R
\name{getReadthrough}
\alias{getReadthrough}
\title{getReadthrough function}
\source{
\href{https://www.gencodegenes.org/human/release_25lift37.html}{Refrence GRCh37 genome}
for details on gtf format visit \href{https://useast.ensembl.org/info/website/upload/gff.html}{ensemble}
}
\usage{
getReadthrough(gr, cod_gr)
}
\arguments{
\item{gr}{A GRanges object that shows the start and end loci on genome.}

\item{cod_gr}{A GRanges object that contains coding regions. For details
please visit /R/data.R.}
}
\value{
Returns the classified transcriptome biotypes.
}
\description{
A getReadthrough functions that will The getReadthrough function is used to
find long transcripts that covers more than two coding regions of a genome.
The variables are as follows.
}
\details{
For details of findOverlaps, type.partialOverlap, type.50Overlap
type.toPlot, queryhits, and subjecthits see,
\href{https://www.bioconductor.org/packages/release/bioc/html/GenomicRanges.html}{GenomicRanges},
\href{https://www.bioconductor.org/packages/release/bioc/html/IRanges.html}{IRanges},
and \href{http://bioconductor.org/install/index.html}{BiocManager}.
}
\note{
Replace the path_file when loading data locally to the data directory.
}
\examples{
data('gencode_gr.v19_chr21.rda')
data('ILEF_gr.chr21.rda')
data('cod_gr.chr21.rda')
getReadthrough(ILEF_gr,cod_gr)

\dontrun{getReadthrough(cod_gr)}

}
\references{
Wang, Z. Z., J. M. Cunningham and X. H. Yang (2018).'CisPi: a transcriptomic score for disclosing cis-acting disease-associated lincRNAs.'
Bioinformatics34(17): 664-670'
}
